use busybox-static from debirf root in nest
Instead of copying busybox and all of it's needed libraries from the
host, we instead just install busybox-static into debirf and then copy
it out from there into the nest.  This is a much simpler operation,
and assures that all the binaries in the nest and in the debirf image
are the same arch, at the expense of a slightly larger image.